move response struct to test folder
