+++ forms/registration_test.go
@@ -2,8 +2,8 @@ package forms_test
 
 import (
 	"go-google-scraper-challenge/forms"
-	. "go-google-scraper-challenge/helpers/test"
 	"go-google-scraper-challenge/initializers"
+	. "go-google-scraper-challenge/test/helpers"
 
 	"github.com/beego/beego/v2/core/validation"
 	. "github.com/onsi/ginkgo"
