+++ models/user_test.go
@@ -1,9 +1,9 @@
 package models_test
 
 import (
-	. "go-google-scraper-challenge/helpers/test"
 	"go-google-scraper-challenge/initializers"
 	"go-google-scraper-challenge/models"
+	. "go-google-scraper-challenge/test/helpers"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
