+++ test/serializers/error.go
@@ -1,0 +1,6 @@
+package tests
+
+type ErrorResponse struct {
+	Error            string `json:"error"`
+	ErrorDescription string `json:"error_description"`
+}
